compiler: make V compilable wit the tcc backend

+++ vlib/builtin/int.v
@@ -31,7 +31,7 @@ pub fn (a f64) eq(b f64) bool {
 	return C.fabs(a - b) <= C.DBL_EPSILON	
 }
 pub fn (a f32) eq(b f32) bool {
-	return C.fabsf(a - b) <= C.FLT_EPSILON	
+	return C.fabs(a - b) <= C.FLT_EPSILON	
 }
 pub fn (a f64) eqbit(b f64) bool {
 	return C.DEFAULT_EQUAL(a, b)
